move NamedProcessCollector to a collector package
This allows the process-exporter to be embedded into applications
directly.
